move physical memtable
