+++ pkg/planner/core/access/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "access",
+    srcs = ["access_obj.go"],
+    importpath = "github.com/pingcap/tidb/pkg/planner/core/access",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_pingcap_tipb//go-tipb"],
+)
