+++ pkg/planner/core/access_object.go
@@ -17,6 +17,7 @@ package core
 import (
 	"context"
 	"fmt"
+	"github.com/pingcap/tidb/pkg/planner/core/access"
 	"sort"
 	"strconv"
 	"strings"
@@ -103,93 +104,6 @@ func (d DynamicPartitionAccessObjects) SetIntoPB(pb *tipb.ExplainOperator) {
 	}
 }
 
-// IndexAccess represents the index accessed by an operator.
-type IndexAccess struct {
-	Name             string
-	Cols             []string
-	IsClusteredIndex bool
-}
-
-// ToPB turns itself into a protobuf message.
-func (a *IndexAccess) ToPB() *tipb.IndexAccess {
-	if a == nil {
-		return nil
-	}
-	return &tipb.IndexAccess{
-		Name:             a.Name,
-		Cols:             a.Cols,
-		IsClusteredIndex: a.IsClusteredIndex,
-	}
-}
-
-// ScanAccessObject represents the access to a table.
-// It may also represent the access to indexes and partitions of a table.
-type ScanAccessObject struct {
-	Database   string
-	Table      string
-	Indexes    []IndexAccess
-	Partitions []string
-}
-
-// NormalizedString implements AccessObject.
-func (s *ScanAccessObject) NormalizedString() string {
-	var b strings.Builder
-	if len(s.Table) > 0 {
-		b.WriteString("table:" + s.Table)
-	}
-	if len(s.Partitions) > 0 {
-		b.WriteString(", partition:?")
-	}
-	for _, index := range s.Indexes {
-		if index.IsClusteredIndex {
-			b.WriteString(", clustered index:")
-		} else {
-			b.WriteString(", index:")
-		}
-		b.WriteString(index.Name + "(" + strings.Join(index.Cols, ", ") + ")")
-	}
-	return b.String()
-}
-
-func (s *ScanAccessObject) String() string {
-	var b strings.Builder
-	if len(s.Table) > 0 {
-		b.WriteString("table:" + s.Table)
-	}
-	if len(s.Partitions) > 0 {
-		b.WriteString(", partition:" + strings.Join(s.Partitions, ","))
-	}
-	for _, index := range s.Indexes {
-		if index.IsClusteredIndex {
-			b.WriteString(", clustered index:")
-		} else {
-			b.WriteString(", index:")
-		}
-		b.WriteString(index.Name + "(" + strings.Join(index.Cols, ", ") + ")")
-	}
-	return b.String()
-}
-
-// SetIntoPB implements AccessObject.
-func (s *ScanAccessObject) SetIntoPB(pb *tipb.ExplainOperator) {
-	if s == nil || pb == nil {
-		return
-	}
-	pbObj := tipb.ScanAccessObject{
-		Database:   s.Database,
-		Table:      s.Table,
-		Partitions: s.Partitions,
-	}
-	for i := range s.Indexes {
-		pbObj.Indexes = append(pbObj.Indexes, s.Indexes[i].ToPB())
-	}
-	pb.AccessObjects = []*tipb.AccessObject{
-		{
-			AccessObject: &tipb.AccessObject_ScanObject{ScanObject: &pbObj},
-		},
-	}
-}
-
 // OtherAccessObject represents other kinds of access.
 type OtherAccessObject string
 
@@ -219,7 +133,7 @@ func (o OtherAccessObject) SetIntoPB(pb *tipb.ExplainOperator) {
 
 // AccessObject implements DataAccesser interface.
 func (p *PhysicalIndexScan) AccessObject() base.AccessObject {
-	res := &ScanAccessObject{
+	res := &access.ScanAccessObject{
 		Database: p.DBName.O,
 	}
 	tblName := p.Table.Name.O
@@ -235,7 +149,7 @@ func (p *PhysicalIndexScan) AccessObject() base.AccessObject {
 		}
 	}
 	if len(p.Index.Columns) > 0 {
-		index := IndexAccess{
+		index := access.IndexAccess{
 			Name: p.Index.Name.O,
 		}
 		for _, idxCol := range p.Index.Columns {
@@ -245,14 +159,14 @@ func (p *PhysicalIndexScan) AccessObject() base.AccessObject {
 				index.Cols = append(index.Cols, idxCol.Name.O)
 			}
 		}
-		res.Indexes = []IndexAccess{index}
+		res.Indexes = []access.IndexAccess{index}
 	}
 	return res
 }
 
 // AccessObject implements DataAccesser interface.
 func (p *PhysicalTableScan) AccessObject() base.AccessObject {
-	res := &ScanAccessObject{
+	res := &access.ScanAccessObject{
 		Database: p.DBName.O,
 	}
 	tblName := p.Table.Name.O
@@ -268,12 +182,12 @@ func (p *PhysicalTableScan) AccessObject() base.AccessObject {
 		}
 	}
 	if len(p.UsedColumnarIndexes) > 0 {
-		res.Indexes = make([]IndexAccess, 0, len(p.UsedColumnarIndexes))
+		res.Indexes = make([]access.IndexAccess, 0, len(p.UsedColumnarIndexes))
 		for _, idx := range p.UsedColumnarIndexes {
 			if idx == nil || idx.IndexInfo == nil {
 				continue
 			}
-			index := IndexAccess{
+			index := access.IndexAccess{
 				Name: idx.IndexInfo.Name.O,
 			}
 			for _, idxCol := range idx.IndexInfo.Columns {
@@ -290,16 +204,8 @@ func (p *PhysicalTableScan) AccessObject() base.AccessObject {
 }
 
 // AccessObject implements DataAccesser interface.
-func (p *PhysicalMemTable) AccessObject() base.AccessObject {
-	return &ScanAccessObject{
-		Database: p.DBName.O,
-		Table:    p.Table.Name.O,
-	}
-}
-
-// AccessObject implements DataAccesser interface.
 func (p *PointGetPlan) AccessObject() base.AccessObject {
-	res := &ScanAccessObject{
+	res := &access.ScanAccessObject{
 		Database: p.dbName,
 		Table:    p.TblInfo.Name.O,
 	}
@@ -314,7 +220,7 @@ func (p *PointGetPlan) AccessObject() base.AccessObject {
 		}
 	}
 	if p.IndexInfo != nil {
-		index := IndexAccess{
+		index := access.IndexAccess{
 			Name:             p.IndexInfo.Name.O,
 			IsClusteredIndex: p.IndexInfo.Primary && p.TblInfo.IsCommonHandle,
 		}
@@ -325,14 +231,14 @@ func (p *PointGetPlan) AccessObject() base.AccessObject {
 				index.Cols = append(index.Cols, idxCol.Name.O)
 			}
 		}
-		res.Indexes = []IndexAccess{index}
+		res.Indexes = []access.IndexAccess{index}
 	}
 	return res
 }
 
 // AccessObject implements DataAccesser interface.
 func (p *BatchPointGetPlan) AccessObject() base.AccessObject {
-	res := &ScanAccessObject{
+	res := &access.ScanAccessObject{
 		Database: p.dbName,
 		Table:    p.TblInfo.Name.O,
 	}
@@ -351,7 +257,7 @@ func (p *BatchPointGetPlan) AccessObject() base.AccessObject {
 		}
 	}
 	if p.IndexInfo != nil {
-		index := IndexAccess{
+		index := access.IndexAccess{
 			Name:             p.IndexInfo.Name.O,
 			IsClusteredIndex: p.IndexInfo.Primary && p.TblInfo.IsCommonHandle,
 		}
@@ -362,7 +268,7 @@ func (p *BatchPointGetPlan) AccessObject() base.AccessObject {
 				index.Cols = append(index.Cols, idxCol.Name.O)
 			}
 		}
-		res.Indexes = []IndexAccess{index}
+		res.Indexes = []access.IndexAccess{index}
 	}
 	return res
 }
