+++ pkg/planner/core/pb_to_plan.go
@@ -120,7 +120,7 @@ func (b *PBPlanBuilder) pbToTableScan(e *tipb.Executor) (base.PhysicalPlan, erro
 		return nil, err
 	}
 	schema := b.buildTableScanSchema(tbl.Meta(), columns)
-	p := PhysicalMemTable{
+	p := physicalop.PhysicalMemTable{
 		DBName:  dbInfo.Name,
 		Table:   tbl.Meta(),
 		Columns: columns,
@@ -307,7 +307,7 @@ func (b *PBPlanBuilder) predicatePushDown(physicalPlan base.PhysicalPlan, predic
 		return predicates, physicalPlan
 	}
 	switch plan := physicalPlan.(type) {
-	case *PhysicalMemTable:
+	case *physicalop.PhysicalMemTable:
 		memTable := plan
 		if memTable.Extractor == nil {
 			return predicates, plan
